feat(product): allow assigned-ids to be used
This commit allows integration flows to assign an ID for categories, products and variants.
This makes it easier  in scenarios where existing services already have assigned a globally known (external) id to the entity.

Example: you are adding medusa to a stack that has an existing recommendation engine in place, that has been using the PIM partnumber as key for a long time. As the PIM is the MDM master, reusing its key, makes it easier for all other systems in the stack to coordinate based on this.

External_id is not searchable or filteriable by default, so it does not serve same function.